+++ common/djangoapps/student/models/__init__.py
@@ -1,0 +1,4 @@
+'''
+Student models migrated to folder to tease out the course enrollment aspects from Student
+'''
+from .student import *
