+++ requirements/edx/base.txt
@@ -50,8 +50,6 @@ babel==2.11.0
     #   enmerkar-underscore
 backoff==1.10.0
     # via analytics-python
-backports-zoneinfo==0.2.1
-    # via icalendar
 beautifulsoup4==4.11.1
     # via pynliner
 billiard==3.6.4.0
@@ -607,10 +605,6 @@ idna==3.4
     #   requests
     #   snowflake-connector-python
     #   yarl
-importlib-metadata==5.0.0
-    # via markdown
-importlib-resources==5.10.0
-    # via jsonschema
 inflection==0.5.1
     # via drf-yasg
 interchange==2021.0.4
@@ -806,8 +800,6 @@ pillow==9.3.0
     #   -r requirements/edx/base.in
     #   edx-enterprise
     #   edx-organizations
-pkgutil-resolve-name==1.3.10
-    # via jsonschema
 polib==1.1.1
     # via edx-i18n-tools
 prompt-toolkit==3.0.31
@@ -1193,10 +1185,6 @@ xss-utils==0.4.0
     # via -r requirements/edx/base.in
 yarl==1.8.1
     # via aiohttp
-zipp==3.10.0
-    # via
-    #   importlib-metadata
-    #   importlib-resources
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
