+++ requirements/edx/development.txt
@@ -79,10 +79,6 @@ backoff==1.10.0
     # via
     #   -r requirements/edx/testing.txt
     #   analytics-python
-backports-zoneinfo==0.2.1
-    # via
-    #   -r requirements/edx/testing.txt
-    #   icalendar
 beautifulsoup4==4.11.1
     # via
     #   -r requirements/edx/testing.txt
@@ -1075,10 +1071,6 @@ pillow==9.3.0
     #   edx-organizations
 pip-tools==6.9.0
     # via -r requirements/edx/pip-tools.txt
-pkgutil-resolve-name==1.3.10
-    # via
-    #   -r requirements/edx/testing.txt
-    #   jsonschema
 platformdirs==2.5.2
     # via
     #   -r requirements/edx/testing.txt
@@ -1580,14 +1572,11 @@ tqdm==4.64.1
 typing-extensions==4.4.0
     # via
     #   -r requirements/edx/testing.txt
-    #   astroid
     #   django-countries
     #   import-linter
     #   mypy
     #   pydantic
-    #   pylint
     #   snowflake-connector-python
-    #   starlette
 unicodecsv==0.14.1
     # via
     #   -r requirements/edx/testing.txt
@@ -1705,11 +1694,6 @@ yarl==1.8.1
     # via
     #   -r requirements/edx/testing.txt
     #   aiohttp
-zipp==3.10.0
-    # via
-    #   -r requirements/edx/testing.txt
-    #   importlib-metadata
-    #   importlib-resources
 
 # The following packages are considered to be unsafe in a requirements file:
 # pip
