+++ requirements/edx/doc.txt
@@ -34,8 +34,6 @@ idna==3.4
     # via requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==5.0.0
-    # via sphinx
 jinja2==3.1.2
     # via
     #   code-annotations
@@ -88,5 +86,3 @@ text-unidecode==1.3
     # via python-slugify
 urllib3==1.26.12
     # via requests
-zipp==3.10.0
-    # via importlib-metadata
