+++ requirements/edx/testing.txt
@@ -73,10 +73,6 @@ backoff==1.10.0
     # via
     #   -r requirements/edx/base.txt
     #   analytics-python
-backports-zoneinfo==0.2.1
-    # via
-    #   -r requirements/edx/base.txt
-    #   icalendar
 beautifulsoup4==4.11.1
     # via
     #   -r requirements/edx/base.txt
@@ -1012,10 +1008,6 @@ pillow==9.3.0
     #   -r requirements/edx/base.txt
     #   edx-enterprise
     #   edx-organizations
-pkgutil-resolve-name==1.3.10
-    # via
-    #   -r requirements/edx/base.txt
-    #   jsonschema
 platformdirs==2.5.2
     # via
     #   pylint
@@ -1464,13 +1456,10 @@ tqdm==4.64.1
 typing-extensions==4.4.0
     # via
     #   -r requirements/edx/base.txt
-    #   astroid
     #   django-countries
     #   import-linter
     #   pydantic
-    #   pylint
     #   snowflake-connector-python
-    #   starlette
 unicodecsv==0.14.1
     # via
     #   -r requirements/edx/base.txt
@@ -1578,11 +1567,6 @@ yarl==1.8.1
     # via
     #   -r requirements/edx/base.txt
     #   aiohttp
-zipp==3.10.0
-    # via
-    #   -r requirements/edx/base.txt
-    #   importlib-metadata
-    #   importlib-resources
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
